+++ src/kwja/cli/utils.py
@@ -91,6 +91,7 @@ def _get_model_version() -> str:
         ("2", "1"): "v2.1",
         ("2", "2"): "v2.2",
         ("2", "3"): "v2.2",
+        ("2", "4"): "v2.4",
     }
     version_fields = kwja.__version__.split(".")
     return version_map[(version_fields[0], version_fields[1])]