+++ src/kwja/utils/constants.py
@@ -16,14 +16,6 @@ MASKED = -1024.0
 RESOURCE_PATH = resource_files(kwja) / "resource"
 
 
-# ---------- word (inference) dataset ----------
-SPLIT_INTO_WORDS_MODEL_NAMES = [
-    "nlp-waseda/roberta-base-japanese",
-    "nlp-waseda/roberta-large-japanese",
-    "nlp-waseda/roberta-large-japanese-seq512",
-]
-
-
 # ---------- typo module ----------
 TYPO_CORR_OP_TAG2TOKEN = {
     "K": "<k>",
@@ -205,47 +197,41 @@ KATA2HIRA = str.maketrans(_KATAKANA, _HIRAGANA)
 # ---------- char module|text normalization ----------
 # 制御文字(\t,\nを含む)は削除
 TRANSLATION_TABLE: Dict[int, Optional[int]] = str.maketrans(
-    ' "#▁', "␣”＃▂", "".join(chr(i) for i in [*range(32), *range(127, 160)])
+    '"#▁', "”＃▂", "".join(chr(i) for i in [*range(32), *range(127, 160), 5760, *range(8232, 8234)])
 )
 
 
 # ---------- seq2seq module----------
-# tokens to separate surface, reading, lemma, and canonical form
 SURF_TOKEN: str = "<extra_id_0>"
 READING_TOKEN: str = "<extra_id_1>"
 LEMMA_TOKEN: str = "<extra_id_2>"
 CANON_TOKEN: str = "<extra_id_3>"
-# tokens to represent full space, half space, no canonical form, and triple dots
-NO_CANON_TOKEN: str = "<extra_id_4>"
-FULL_SPACE_TOKEN: str = "<extra_id_5>"
-HALF_SPACE_TOKEN1: str = "<extra_id_6>"
-HALF_SPACE_TOKEN2: str = "<extra_id_7>"
-TRIPLE_DOT_TOKEN: str = "<extra_id_8>"
-# token to split input text into morphemes
-MORPHEME_SPLIT_TOKEN: str = "<extra_id_9>"
-# tokens for unk tokens
-RARE_TO_SPECIAL: Dict[str, str] = {
-    "ゔ": "<extra_id_10>",
-    "榕": "<extra_id_11>",
-    "謄": "<extra_id_12>",
-    "丿": "<extra_id_13>",
-    "孜": "<extra_id_14>",
-    "腑": "<extra_id_15>",
-    "庖": "<extra_id_16>",
-    "┘": "<extra_id_17>",
-    "秧": "<extra_id_18>",
-    "褪": "<extra_id_19>",
-    "疥": "<extra_id_20>",
-    "鮪": "<extra_id_21>",
-    "髑髏": "<extra_id_22>",
-    "侭": "<extra_id_23>",
-    "蒟蒻": "<extra_id_24>",
-    "╹": "<extra_id_25>",
-    "厂": "<extra_id_26>",
-    "Ӧ": "<extra_id_27>",
-    "溢": "<extra_id_28>",
+NO_CANON_TOKEN: str = "<extra_id_4>"  # control token to represent no canonical form
+MORPHEME_DELIMITER_TOKEN: str = "<extra_id_5>"  # control token to segment input text into morphemes
+# special tokens
+HALF_SPACE_TOKEN: str = "<extra_id_6>"
+RARE2SPECIAL: Dict[str, str] = {
+    "ゔ": "<extra_id_7>",
+    "榕": "<extra_id_8>",
+    "謄": "<extra_id_9>",
+    "丿": "<extra_id_10>",
+    "孜": "<extra_id_11>",
+    "腑": "<extra_id_12>",
+    "庖": "<extra_id_13>",
+    "┘": "<extra_id_14>",
+    "秧": "<extra_id_15>",
+    "褪": "<extra_id_16>",
+    "疥": "<extra_id_17>",
+    "鮪": "<extra_id_18>",
+    "髑髏": "<extra_id_19>",
+    "侭": "<extra_id_20>",
+    "蒟蒻": "<extra_id_21>",
+    "╹": "<extra_id_22>",
+    "厂": "<extra_id_23>",
+    "Ӧ": "<extra_id_24>",
+    "溢": "<extra_id_25>",
 }
-SPECIAL_TO_RARE: Dict[str, str] = {v: k for k, v in RARE_TO_SPECIAL.items()}
+SPECIAL2RARE: Dict[str, str] = {v: k for k, v in RARE2SPECIAL.items()}
 
 
 # ---------- word module ----------
@@ -1054,7 +1040,15 @@ CONJTYPE_TAG_CONJFORM_TAG2CONJFORM_ID: Dict[str, Dict[str, int]] = {
         "ヤ列基本推量形": 11,
         "ヤ列基本省略推量形": 12,
     },
-    "助動詞そうだ型": {"*": 0, "語幹": 1, "基本形": 2, "ダ列タ系連用テ形": 3, "デアル列基本形": 4, "デス列基本形": 5, "デス列音便基本形": 6},
+    "助動詞そうだ型": {
+        "*": 0,
+        "語幹": 1,
+        "基本形": 2,
+        "ダ列タ系連用テ形": 3,
+        "デアル列基本形": 4,
+        "デス列基本形": 5,
+        "デス列音便基本形": 6,
+    },
     "助動詞く型": {"*": 0, "語幹": 1, "基本形": 2, "基本連用形": 3, "文語連体形": 4, "文語未然形": 5},
     "動詞性接尾辞ます型": {
         "*": 0,